# perf: optimize Docker image