--- a/.dockerignore
+++ b/.dockerignore
@@ -1,7 +1,10 @@
 .DS_Store
+**/__pycache__
+**/*.py[cod]
 .gitignore
 .git
 .github
 .mypy_cache
 .venv
 .tox
+dist