--- a/.dockerignore
+++ b/.dockerignore
@@ -7,4 +7,3 @@
 .mypy_cache
 .venv
 .tox
-dist