--- a/.github/workflows/release.yml
+++ b/.github/workflows/release.yml
@@ -42,6 +42,8 @@ jobs:
     defaults:
       run:
         shell: bash
+    env:
+      DOCKER_BUILDKIT: 1
     steps:
       - name: Checkout the repo
         uses: actions/checkout@v3
@@ -151,7 +153,7 @@ jobs:
         run: |
           export PKG_SRC=dist/phylum-*.whl
           export PKG_NAME=phylum-*.whl
-          docker build --tag phylum-ci --build-arg PKG_SRC --build-arg PKG_NAME .
+          docker build --tag phylum-ci --build-arg PKG_SRC --build-arg PKG_NAME --build-arg BUILDKIT_INLINE_CACHE=1 .
           docker tag phylum-ci phylumio/phylum-ci:latest
           docker tag phylum-ci phylumio/phylum-ci:${{ env.phylum_release_version }}