--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -81,6 +81,8 @@ jobs:
     defaults:
       run:
         shell: bash
+    env:
+      DOCKER_BUILDKIT: 1
     steps:
       - name: Checkout the repo
         uses: actions/checkout@v3