Some of the changes made to complete this issue and resulting improvements:

* The image base layer was switched from one based on Debian to one based on Alpine
* The compressed image size was reduced from 84MB to 35MB
* The expanded image size was reduced from 245MB to 96MB
* The image acquisition process in GitLab was reduced from ~16s to ~9s and the total run time was reduced from ~26s to ~19s
* The Dockerfile now makes use of BuildKit for caching purposes
  * It can be used to create images used for layer caching
  * It is also making use of a cache type build mount in order to re-use already built wheels
    * This cuts the build time for cross compiled builds by ~75% (after the wheels are in the cache)

Additional changes made in this PR, unrelated to issue #48:

* Reformat code examples to add whitespace lines
* Expose version arguments with a short form `-V`
  * The lowercase form `-v` is reserved for future use, to represent the short form of `--verbose`

Closes #48

## Checklist

- [x] Does this PR have an associated issue (i.e., `closes #<issueNum>` in description above)?
- [x] Have you ensured that you have met the expected acceptance criteria?
- [ ] ~Have you created sufficient tests?~
  - Still only manually tested
- [x] Have you updated all affected documentation?

## Screenshots

---

GitLab run before the changes:

<img width="2285" alt="image" src="https://user-images.githubusercontent.com/18729796/172653281-154fb119-11d8-4069-a55b-551d098c8ebe.png">

---

GitLab run after the changes:

<img width="2285" alt="image" src="https://user-images.githubusercontent.com/18729796/172653523-33fc2421-ea09-4461-b1e4-12e87c4f73e6.png">